==> code/01-wsl-scraper.ipynb <==
# World Surf League Scores Web Scraping

## Imports

import pandas as pd
import numpy as np
from requests import get
import json
from bs4 import BeautifulSoup
import time



## World Surf League Scores

I'm going to focus on the Pipeline Masters first. It is the marquee event of the Championship Tour and is run every year (**verify**).

### Get HTML soup and review

Use Beautiful Soup to get html from World Surf League event pages. Each page has the final result and a 'carousel' that links to previous rounds.

# Starting URLs
pipe_2008_url = "https://www.worldsurfleague.com/events/2008/mct/75/billabong-pipeline-masters"
res = get(pipe_2008_url)
soup = BeautifulSoup(res.content)


From inspecting page source I have found an unordered list with useful urls to get all scores from each event

# Find all list items within the event rounds div
carousel = soup.find('div', id='event-rounds').find_all('li')

# Visual check of event round urls and round names/numbers
for horse in carousel:
    link_elements = horse.find_all('a')
    print(link_elements)

# Build a dict of urls and round names
round_urls = []
for horse in carousel:
    link_elements = horse.find_all('a')
    for link in link_elements:
        url_dict = {}
        url_dict['round'] = link.text
        # Add endpoint for ease of use later
        url_dict['url'] = "https://www.worldsurfleague.com" + link['href']
        print(link['href'], link.text)
        round_urls.append(url_dict)

# With help from https://stackoverflow.com/questions/11092511/list-of-unique-dictionaries
# Keep only unique values
round_urls_unique = list({unique['url']:unique for unique in round_urls}.values())
round_urls_unique

### Get scores for every round in an event

# Get scores function takes the url dict and gets scores for each round of the event
def get_scores(url_list_of_dicts, event_year, event_name):
    data = []
    # Iterate through event rounds
    for round in url_list_of_dicts:
        # Get html soup
        res = get(round["url"])
        round_soup = BeautifulSoup(res.content)
        # Get list of athlete names
        athletes = round_soup.find_all("span", class_="athlete-name")

        # Get list of scores lists
        scores = round_soup.find_all("div", class_="all-waves all-waves-grid")

        # Go through athletes
        for i in range(len(athletes)):

            # Create dict to save scores
            athlete_scores = {}

            # Save event name and year
            athlete_scores["year"] = event_year
            athlete_scores["event"] = event_name

            # Save round name
            athlete_scores["round"] = round["round"]

            # Save heat number
            athlete_scores["heat"] = (
                athletes[i]
                .find_parent("div", class_="bd new-heat-bd")
                .find_previous_sibling()
                .find("span", class_="new-heat-hd-name")
                .text
            )

            # Athlete name
            athlete_scores["name"] = athletes[i].text

            # Empty list for athletes scores in that heat
            score_list = []

            # Scores list
            per_wave_scores = scores[i].find_all("span", class_="score")

            # Append scores
            for b in range(len(per_wave_scores)):
                try:
                    score_list.append(float(per_wave_scores[b].text))
                except:
                    continue

            # Get two top scores of each athlete per heat
            if len(score_list) > 0:
                top_score = max(score_list)
                score_list.remove(top_score)
                if len(score_list) > 0:
                    second_score = max(score_list)
                    # Assign total score to data
                    athlete_scores["top_two_waves_total"] = top_score + second_score
                else:
                    athlete_scores["top_two_waves_total"] = top_score
            else:
                # Assign total score to data
                athlete_scores["top_two_waves_total"] = 0.0

            # Add to data list
            data.append(athlete_scores)

        # Sleep
        time.sleep(1)

    return data


pipe_2008_scores = get_scores(round_urls_unique, '2008', 'pipe')

pd.DataFrame(pipe_2008_scores)

### Get all scores for all Pipeline events in all years

# Get urls for each year's events page
yearly_urls = []
for year in range(2008, 2023):
    # Exclude 2020, all events were cancelled
    if year != 2020:
        dict_list = {}
        dict_list['year'] = year
        dict_list['url'] = f'https://www.worldsurfleague.com/events/{year}/mct?all=1'
        yearly_urls.append(dict_list)

# Get URL for each event accross years
event_urls = []
for year in yearly_urls:
    res = get(year['url'])
    soup = BeautifulSoup(res.content)
    for each in soup.find_all("a", class_="event-schedule-details__event-name"):
        # Only get pipeline events
        if 'Pipe' in each.text.replace('/',''):
            event = {}
            event['year'] = year['year']
            event['name'] = each.text.replace('/','')
            event['url'] = each['href']
            event_urls.append(event)
        else:
            continue
    time.sleep(1)

event_urls

# Get all scores and save them as a list of df
df_list = []
for event in event_urls:
    print(f"Getting {event['year']} {event['name']}...")
    res = get(event["url"])
    soup = BeautifulSoup(res.content)
    # Find all list items within the event rounds div
    try:
        carousel = soup.find("div", id="event-rounds").find_all("li")
    except:
        print(f"FAILED {event['year']} {event['name']}")
        continue

    # Build a dict of urls and round names
    round_urls = []
    for horse in carousel:
        link_elements = horse.find_all("a")
        for link in link_elements:
            url_dict = {}
            url_dict["round"] = link.text
            # Add endpoint for ease of use later
            url_dict["url"] = "https://www.worldsurfleague.com" + link["href"]
            round_urls.append(url_dict)

    # Keep only unique values
    round_urls_unique = list({unique["url"]: unique for unique in round_urls}.values())

    # Run get scores function and assign result to a variable
    event_scores = get_scores(
        round_urls_unique, event_year=event["year"], event_name=event["name"]
    )

    # Convert scores data to DataFrame and append to df_list
    # pd.DataFrame(event_scores).to_csv(
    #     f"./data/{event['name']} {event['year']}.csv", index=False
    # )
    df_list.append(pd.DataFrame(event_scores))
    print(f"... got {event['year']} {event['name']}")


pd.concat(df_list)

pd.concat(df_list).to_csv('../data/wsl/pipeline_avg_scores_all_years.csv')

## World Surf League Dates and Times of Rounds

Using a similar loop as before, get dates for event heats where possible

res = get(event_urls[-4]["url"] + "/results")
soup = BeautifulSoup(res.content)
results_grid = soup.find(
    "div", class_="post-event-watch-heat-grid__body post-event-watch-module-body"
)
round_name = results_grid.find(
    "div", class_="post-event-watch-heat-grid__round-header"
).text.strip()
heats = results_grid.find_all("div", class_="hot-heat__hd")


heats[2].find("span", class_="heat-name").text


heats[2].find("span", class_="hot-heat__status hot-heat__status--over").text


# Find carousel and get round urls for each event
carousel = soup.find("div", class_="post-event-watch-round-nav post-event-watch-module")
links = carousel.find_all("a")
for link in links:
    print(link['href'])


# Build a data frame with year,event,round,heat,date columns for all events all years
df_list = []
for event in event_urls:
    data = []
    print(f"Getting {event['year']} {event['name']}...")
    res = get(event["url"] + "/results")
    soup = BeautifulSoup(res.content)
    carousel = soup.find(
        "div", class_="post-event-watch-round-nav post-event-watch-module"
    )
    links = carousel.find_all("a")
    for link in links:
        endpoint = link["href"].split("?")[1]
        res2 = get(event["url"] + "/results?" + endpoint)
        soup2 = BeautifulSoup(res2.content)
        results_grid = soup2.find(
            "div",
            class_="post-event-watch-heat-grid__body post-event-watch-module-body",
        )

        if results_grid:
            round_name = results_grid.find(
                "div", class_="post-event-watch-heat-grid__round-header"
            ).text.strip()
            heats = results_grid.find_all("div", class_="hot-heat__hd")
            for heat in heats:
                heat_data = {}
                try:
                    heat_data["year"] = event["year"]
                    heat_data["event"] = event["name"]
                    heat_data["round"] = round_name
                    try:
                        heat_data["heat"] = heat.find("span", class_="heat-name").text
                    except:
                        heat_data["heat"] = 'Heat 1'
                    heat_data["date"] = heat.find(
                        "span", class_="hot-heat__status hot-heat__status--over"
                    ).text.replace('Ended', '')
                    data.append(heat_data)
                    print(heat_data['heat'], heat_data['date'])
                except:
                    print(
                        f"something wrong in heat data for {event['year']} {event['name']}"
                    )
                    continue
        else:
            print(f"FAILED {event['year']} {event['name']}")
            continue
        time.sleep(1)
    print(f"... got {event['year']} {event['name']}")

    df_list.append(pd.DataFrame(data))
    time.sleep(2)


dates = pd.concat(df_list)
dates[dates['year'] == 2022]['date'].value_counts()

dates[dates['year'] == 2022]['round'].value_counts()

dates['round'].value_counts()

dates['heat'].value_counts()

dates.to_csv('../data/wsl/pipe_masters_heat_dates.csv')

## All events all years

### Get all scores for all events in all years

# Get urls for each year's events page
yearly_urls = []
for year in range(2008, 2023):
    # Exclude 2020, all events were cancelled
    if year != 2020:
        dict_list = {}
        dict_list['year'] = year
        dict_list['url'] = f'https://www.worldsurfleague.com/events/{year}/mct?all=1'
        yearly_urls.append(dict_list)

# Get URL for each event accross years
event_urls = []
for year in yearly_urls:
    res = get(year['url'])
    soup = BeautifulSoup(res.content)
    for each in soup.find_all("a", class_="event-schedule-details__event-name"):
        event = {}
        event['year'] = year['year']
        event['name'] = each.text.replace('/','')
        event['url'] = each['href']
        event_urls.append(event)
        
    time.sleep(1)

event_urls

# Get all scores and save them as a list of df
df_list = []
for event in event_urls:
    print(f"Getting {event['year']} {event['name']}...")
    res = get(event["url"])
    soup = BeautifulSoup(res.content)
    # Find all list items within the event rounds div
    try:
        carousel = soup.find("div", id="event-rounds").find_all("li")
    except:
        print(f"FAILED {event['year']} {event['name']}")
        continue

    # Build a dict of urls and round names
    round_urls = []
    for horse in carousel:
        link_elements = horse.find_all("a")
        for link in link_elements:
            url_dict = {}
            url_dict["round"] = link.text
            # Add endpoint for ease of use later
            url_dict["url"] = "https://www.worldsurfleague.com" + link["href"]
            round_urls.append(url_dict)

    # Keep only unique values
    round_urls_unique = list({unique["url"]: unique for unique in round_urls}.values())

    # Run get scores function and assign result to a variable
    event_scores = get_scores(
        round_urls_unique, event_year=event["year"], event_name=event["name"]
    )

    
    # pd.DataFrame(event_scores).to_csv(
    #     f"./data/{event['name']} {event['year']}.csv", index=False
    # )

    # Convert scores data to DataFrame and append to df_list
    df_list.append(pd.DataFrame(event_scores))
    print(f"... got {event['year']} {event['name']}")


pd.concat(df_list)

pd.concat(df_list).to_csv('../data/wsl/avg_scores_all_years.csv')
