==> code/04-clean-and-merge-all-data.ipynb <==
# WSL Merge

Merge score data with date and time data

## Imports

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt


## Scores Data

df = pd.read_csv("../data/wsl/avg_scores_all_years.csv", index_col=0).rename(columns={'top_two_waves_total': 'score'})
df.head()


Look at unique events and find total amoun of heats for the Pipeline events. Event names change with sponsorship changes.

df["event"].unique()


Locate rows with pipeline events and get the shape of the resulting dataframe. This is the amount of heats in pipeline events with data.

# Make list of pipeline events from previous cell
pipe = [
    "Billabong Pipe Masters presented by Hydro Flask",
    "Billabong Pro Pipeline",
    "Billabong Pipe Masters presented by Hydro Flask",
    "Billabong Pipe Masters",
    "Billabong Pipeline Masters",
]

df.loc[df["event"].isin(pipe)].shape


Check all other events to confirm pipeline has the most data.

df["event"].value_counts().head(15)


**Pipeline is the event with most data**

Trestles is another good target with data and historical buoy info.

# Cast year data as string to avoid groupby issues
df["year"] = df["year"].astype(str)


### Check and rename rounds and heats for consistency accross events

I've mapped these manually but str methods would reduce the amount of code

# Get unique round names
df["round"].unique()


# Map round names to a consistent standard
# Round 1, 2, 3, 4 and 5, then Quarterfinals, Semifinals and Final. Idem for trials
df["round"] = df["round"].map(
    {
        "Seeding Round": "Round 1",
        "Round of 32": "Round 3",
        "Quarterfinal": "Quarterfinals",
        "Round of 16": "Round 4",
        "Elimination Round": "Round 2",
        "Opening Round": "Round 1",
        "Round 1": "Round 1",
        "Round 2": "Round 2",
        "Round 3": "Round 3",
        "Round 4": "Round 4",
        "Round 5": "Round 5",
        "Quarterfinals": "Quarterfinals",
        "Final": "Final",
        "Semifinal": "Semifinals",
        "Semifinals": "Semifinals",
        "Trials Quarterfinals": "Trials Quarterfinals",
        "Trials Round 1": "Trials Round 1",
        "Trials Semifinals": "Trials Semifinals",
        "Trials Final": "Trials Final",
    }
)


# Get unique heat names
df['heat'].unique()

# Map for consistency and to eliminate 'Watch replay' strings from webscraping
df["heat"] = df["heat"].map(
    {
        "Heat 1 ": "Heat 1",
        "Heat 2 ": "Heat 2",
        "Heat 3 ": "Heat 3",
        "Heat 4 ": "Heat 4",
        "Heat 5 ": "Heat 5",
        "Heat 6 ": "Heat 6",
        "Heat 7 ": "Heat 7",
        "Heat 8 ": "Heat 8",
        "Heat 9 ": "Heat 9",
        "Heat 10 ": "Heat 10",
        "Heat 11 ": "Heat 11",
        "Heat 12 ": "Heat 12",
        "Heat 13 ": "Heat 13",
        "Heat 14 ": "Heat 14",
        "Heat 15 ": "Heat 15",
        "Heat 16 ": "Heat 16",
        "Heat 8 Watch replay": "Heat 8",
        "Heat 6 Watch replay": "Heat 6",
        "Heat 2 Watch replay": "Heat 2",
        "Heat 1 Watch replay": "Heat 1",
        "Heat 4 Watch replay": "Heat 4",
        "Heat 10 Watch replay": "Heat 10",
        "Heat 11 Watch replay": "Heat 11",
        "Heat 12 Watch replay": "Heat 12",
        "Heat 3 Watch replay": "Heat 3",
        "Heat 5 Watch replay": "Heat 5",
        "Heat 7 Watch replay": "Heat 7",
        "Heat 9 Watch replay": "Heat 9",
        "Heat 13 Watch replay": "Heat 13",
        "Heat 14 Watch replay": "Heat 14",
        "Heat 15 Watch replay": "Heat 15",
        "Heat 16 Watch replay": "Heat 16",
    }
)


## Save for EDA

df.to_csv('../data/wsl/avg_all_years_eda.csv')

## Focus on Pipeline data

**Reindex doesn't work as expected**

# Get only pipeline scores
# Uses 'pipe' list from earlier
df_pipe = df.loc[df["event"].isin(pipe)].reindex()
df_pipe


## Load event dates data

This data was scraped idependently from scores to avoid using selenium

# Pipeline heat dates scraped from /results endpoint instead of main
event_dates = pd.read_csv("../data/wsl/pipe_masters_heat_dates.csv", index_col=0)
event_dates.head()


### Missing data

Events before 2014 have no heat dates. Have contacted WSL to inquire about getting this data in a structured way.

Other option would be to investigate news articles manually but I'll do that if I have time after project is completed

event_dates["date"].value_counts()


# Sanity check
event_dates[event_dates["year"] == 2022]["round"].value_counts()


### Round and heat names have to be mapped like score data

event_dates["round"].unique()


df_pipe["round"].unique()


event_dates["round"] = event_dates["round"].map(
    {
        "Seeding Round": "Round 1",
        "Round of 32": "Round 3",
        "Quarterfinal": "Quarterfinals",
        "Round of 16": "Round 4",
        "Elimination Round": "Round 2",
        "Opening Round": "Round 1",
        "Round 1": "Round 1",
        "Round 2": "Round 2",
        "Round 3": "Round 3",
        "Round 4": "Round 4",
        "Round 5": "Round 5",
        "Quarterfinals": "Quarterfinals",
        "Final": "Final",
        "Semifinal": "Semifinals",
        "Semifinals": "Semifinals",
        "Trials Quarterfinals": "Trials Quarterfinals",
        "Trials Round 1": "Trials Round 1",
        "Trials Semifinals": "Trials Semifinals",
        "Trials Final": "Trials Final",
    }
)


event_dates["heat"].unique()


df_pipe["heat"].unique()


df_pipe["heat"].unique() == event_dates["heat"].unique()


df_pipe["round"].unique() == event_dates["round"].unique()


### Convert dates to datetime

# Check unique values
event_dates["date"].unique()


# Eliminate first character blank space
event_dates["date"] = event_dates["date"].str[1:]


Keep only rows with date data

event_dates = event_dates[event_dates["date"] != ""]


event_dates["date"] = pd.to_datetime(event_dates["date"])
event_dates["date"].unique()


Save to csv and assign offsets in Excel

event_dates.to_csv("../data/wsl/pipeline_dates.csv")


Read in excel with offsets. This could be accomplished with a .apply function

event_dates_offset = pd.read_csv("../data/wsl/pipeline_dates_offset.csv", index_col=0)
# Convert date column to datetime format
event_dates_offset["date"] = pd.to_datetime(event_dates_offset["date"])
event_dates_offset['year'] = event_dates_offset['year'].astype(str)
event_dates_offset.head()


### Assign estimated times to heats based on dates and 30 minute heat times

# With help from
# https://www.codegrepper.com/code-examples/python/how+to+add+30+minutes+in+datetime+column+in+pandas
event_dates_offset["est_datetime"] = event_dates_offset.apply(
    lambda x: x["date"] + pd.DateOffset(hours=x["hour"]),
    axis=1,
)


event_dates_offset.head()


event_dates_offset.tail()

## Merge scores with times

scores_and_times = df_pipe.merge(
    event_dates_offset,
    left_on=["year", "event", "round", "heat"],
    right_on=["year", "event", "round", "heat"],
).drop(columns=["date", "hour"])
scores_and_times.head(30)


scores_cleaned_merged = pd.DataFrame(scores_and_times.groupby(by=['est_datetime']).mean())
scores_cleaned_merged

## Save scores and times to csv

scores_cleaned_merged.to_csv('../data/wsl/pipe_2014_2022_clean_dated.csv')

## Get buoy data, wind data; and merge with scores and times

noaa = pd.read_csv('../data/noaa/pipeline_buoy_full_shifted.csv', index_col=0, parse_dates=True)
noaa

wind = pd.read_csv('../data/wind/oahu_wind.csv', index_col=0, parse_dates=True)
wind

data = noaa.merge(
    scores_cleaned_merged.merge(wind, how="left", left_index=True, right_index=True),
    left_index=True,
    right_index=True,
)
data.head()


# Sanity check
data.info()

## Save EDA and Modeling data

data.to_csv('../data/merged_data.csv')

## Linear Model fun

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LinearRegression, Lasso, Ridge
from sklearn.model_selection import GridSearchCV, train_test_split
from sklearn.metrics import mean_squared_error

X = data.drop(
    columns=[
        "score",
    ]
)
y = data['score']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=1331)

ss = StandardScaler()
X_train_sc = ss.fit_transform(X_train)
X_test_sc = ss.fit_transform(X_test)

linreg = LinearRegression()
linreg.fit(X_train_sc, y_train)
print(linreg.score(X_test_sc, y_test))

pd.DataFrame(linreg.coef_, X.columns).sort_values(by=0, ascending=False)

mean_squared_error(y_test, linreg.predict(X_test_sc), squared=False)
