==> code/06-linear-models.ipynb <==
# Linear Models

Build a baseline a linear regression models to predict scores.

RMSE is tha main metric for model evaluation.

## Imports

import pandas as pd
import numpy as np

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LinearRegression, Lasso, Ridge
from sklearn.model_selection import GridSearchCV, train_test_split
from sklearn.metrics import mean_squared_error

data = pd.read_csv('../data/merged_data.csv')
data.drop(columns='Unnamed: 0', inplace=True)
data

## X, y

# split data into input and output columns
X = data.drop(columns="score")
y = data["score"]

X.columns

## Standard Scaling

ss = StandardScaler()
X_sc = ss.fit_transform(X)

## Train, test split

# train test split
X_train, X_test, y_train, y_test = train_test_split(X_sc, y, random_state=1331)

## Baseline

y_baseline_preds = np.full_like(y, y.mean()) 
baseline_rmse = mean_squared_error(y, y_baseline_preds, squared=False)
baseline_rmse

## Create list of dicts to save model results

all_model_results = []


### Add baseline score

baseline_results = {
    "Model Name": "Baseline (y_mean)",
    "Score Type": 'Baseline RMSE',
    "Score": baseline_rmse,
    "Observations": "",
}
all_model_results.append(baseline_results)
all_model_results

## Linear Regression

# Instantiate model
linreg = LinearRegression()

# Fit on training set
linreg.fit(X_train, y_train)

# Score on training and testing set
print(f"R-squared score on train set: {linreg.score(X_train, y_train)}")
print(f"R-squared score on test set: {linreg.score(X_test, y_test)}")

# Review strongest coefficients
pd.DataFrame(linreg.coef_, X.columns).sort_values(by=0, ascending=False).head()

# Review strongest coefficients
pd.DataFrame(linreg.coef_, X.columns).sort_values(by=0, ascending=False).tail()

### Linear Regression RMSE

linreg_train_rmse = mean_squared_error(y_train, linreg.predict(X_train), squared=False)
linreg_train_rmse

linreg_test_rmse = mean_squared_error(y_test, linreg.predict(X_test), squared=False)
linreg_test_rmse

### Save model result for final comparisson

linreg_results = [
    {
        "Model Name": "Linear Regression",
        "Score Type": "Train RMSE",
        "Score": linreg_train_rmse,
        "Observations": "",
    },
    {
        "Model Name": "Linear Regression",
        "Score Type": "Test RMSE",
        "Score": linreg_test_rmse,
        "Observations": "",
    },
]

for result in linreg_results:
    all_model_results.append(result)


## Lasso Regression with Grid Search CV

# Empty dict for results
lasso_results = {}

# Instantiate model
lasso = Lasso()

# Define GS params
lasso_params = {"alpha": [0.01, 0.1, 1, 10, 100]}

# Instantiate GridSearch
lasso_gs = GridSearchCV(lasso, lasso_params)

# Fit on training set
lasso_gs.fit(X_train, y_train)

# Score on training and testing set
print(f"R-squared score on train set: {lasso_gs.score(X_train, y_train)}")
print(f"R-squared score on test set: {lasso_gs.score(X_test, y_test)}")


# Get Lasso best alpha
lasso_alpha = lasso_gs.best_params_
lasso_alpha

# Review coefficients
lasso_coef = pd.DataFrame(lasso_gs.best_estimator_.coef_, X.columns).sort_values(by=0, ascending=False)
lasso_coef = lasso_coef[lasso_coef[0] != 0]
lasso_coef

### Lasso Regression RMSE

lasso_train_rmse = mean_squared_error(y_train, lasso_gs.best_estimator_.predict(X_train), squared=False)
lasso_train_rmse

lasso_test_rmse = mean_squared_error(y_test, lasso_gs.best_estimator_.predict(X_test), squared=False)
lasso_test_rmse

### Save model result for final comparisson

lasso_results = [
    {
        "Model Name": "Lasso Regression",
        "Score Type": "Train RMSE",
        "Score": lasso_train_rmse,
        "Observations": lasso_alpha,
    },
    {
        "Model Name": "Lasso Regression",
        "Score Type": "Test RMSE",
        "Score": lasso_test_rmse,
        "Observations": lasso_alpha,
    },
]

for result in lasso_results:
    all_model_results.append(result)


### Save model pkl for streamlit

# With help from 6.07 lesson
import pickle
with open("../saved-models/saved-lasso.pkl", "wb") as file:
    pickle.dump(lasso_gs.best_estimator_, file)

## Ridge Regression with Grid Search CV

# Empty dict for results
ridge_results = {}

# Instantiate model
ridge = Ridge()

# Define GS params
ridge_params = {'alpha': [100, 200, 500, 1_000]}

# Instantiate GridSearch
ridge_gs = GridSearchCV(ridge, ridge_params)

# Fit on training set
ridge_gs.fit(X_train, y_train)

# Score on training and testing set
print(f"R-squared score on train set: {ridge_gs.score(X_train, y_train)}")
print(f"R-squared score on test set: {ridge_gs.score(X_test, y_test)}")




# Get Ridge best alpha
ridge_alpha = ridge_gs.best_params_
ridge_alpha

# Review coefficients
ridge_coef = pd.DataFrame(ridge_gs.best_estimator_.coef_, X.columns).sort_values(by=0, ascending=False)
ridge_coef = ridge_coef[ridge_coef[0] != 0]
ridge_coef

### Ridge Regression RMSE

ridge_train_rmse = mean_squared_error(y_train, ridge_gs.best_estimator_.predict(X_train), squared=False)
ridge_train_rmse

ridge_test_rmse = mean_squared_error(y_test, ridge_gs.best_estimator_.predict(X_test), squared=False)
ridge_test_rmse

### Save model result for final comparisson

ridge_results = [
    {
        "Model Name": "Ridge Regression",
        "Score Type": "Train RMSE",
        "Score": ridge_train_rmse,
        "Observations": ridge_alpha,
    },
    {
        "Model Name": "Ridge Regression",
        "Score Type": "Test RMSE",
        "Score": ridge_test_rmse,
        "Observations": ridge_alpha,
    },
]

for result in ridge_results:
    all_model_results.append(result)
all_model_results

# Linear Regression Results

pd.DataFrame(all_model_results)

## Save results to csv

pd.DataFrame(all_model_results).to_csv('../model-results/linreg.csv')
