==> code/07-xgboost-models.ipynb <==
# XGBoost model

## Imports

# With help from https://machinelearningmastery.com/xgboost-for-regression/

import numpy as np
import pandas as pd
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import cross_val_score, train_test_split
from sklearn.model_selection import RepeatedKFold, GridSearchCV
from sklearn.metrics import mean_squared_error
from xgboost import XGBRegressor


from google.colab import drive
drive.mount('/content/drive')

### Data

# # local data

# data = pd.read_csv('../data/merged_data.csv')
# data.drop(columns='Unnamed: 0', inplace=True)
# data.head()


# Collab data

data = pd.read_csv('/content/drive/Othercomputers/My MacBook Air/Dropbox/ga/capstone/data/merged_data.csv')
data.drop(columns='Unnamed: 0', inplace=True)
data.head()


## X, y

# split data into input and output columns
X = data.drop(columns="score")
y = data["score"]

## Standard Scaling

ss = StandardScaler()
X_sc = ss.fit_transform(X)

## Train, test split

# train test split
X_train, X_test, y_train, y_test = train_test_split(X_sc, y, random_state=1331)

## Baseline Model RSME

y_baseline_preds = np.full_like(y, y.mean()) 
mean_squared_error(y, y_baseline_preds, squared=False)

## Create list of dicts to save model results

all_model_results = []

## XGBoost model, GridSearchCV

Grid search over:
- n_estimators
- max_depth
- learning_rate

# Empty dict for results
xgb_results = {}

# Instantiate model
model = XGBRegressor(objective='reg:squarederror')

# Define GS params
xgb_params = {'max_depth': [1, 3, 5],
    'n_estimators': [2_000, 5_000, 10_000],
    'learning_rate': [.1, .01, .001, .0001] 
}

# Instantiate GridSearch
xgb_gs = GridSearchCV(model, xgb_params)

# Fit on training set
xgb_gs.fit(X_train, y_train)

# Score on training and testing set
print(f"R-squared score on train set: {xgb_gs.score(X_train, y_train)}")
print(f"R-squared score on test set: {xgb_gs.score(X_test, y_test)}")


# Get best params
xgb_best_params = xgb_gs.best_params_
xgb_best_params

### XGBoost Regression RMSE

xgb_train_rmse = mean_squared_error(y_train, xgb_gs.best_estimator_.predict(X_train), squared=False)
xgb_train_rmse


xgb_test_rmse = mean_squared_error(y_test, xgb_gs.best_estimator_.predict(X_test), squared=False)
xgb_test_rmse


### Save model result for final comparisson

xgb_results = [
    {
        "Model Name": "XGBoost Regression",
        "Score Type": "Train RMSE",
        "Score": xgb_train_rmse,
        "Observations": xgb_best_params,
    },
    {
        "Model Name": "XGBoost Regression",
        "Score Type": "Test RMSE",
        "Score": xgb_test_rmse,
        "Observations": xgb_best_params,
    },
]

for result in xgb_results:
    all_model_results.append(result)
all_model_results

## Save results to csv

# pd.DataFrame(all_model_results).to_csv('../model-results/xgboost.csv')

pd.DataFrame(all_model_results).to_csv('/content/drive/Othercomputers/My MacBook Air/Dropbox/ga/capstone/model-results/xgboost.csv')
